refactor tests for forms
